+++ src/Calendar/Container.scss
@@ -1,4 +1,4 @@
-@import "./variables";
+@import "../variables";
 
 .root {
   position: relative;
